Moved numerics parsing extensions to StringExtensions type
